Adaptive Heart health system, Game loader, Fixed bugs

+++ Assets/Scripts/AttackSystem/AttackSystem.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public abstract class AttackSystem : MonoBehaviour
 {
-    public abstract void Attack(float damage, float weaponLength, Weapon weapon);
+    public abstract void Attack(int damage, float weaponLength, Weapon weapon);
     public abstract void Block();
 }
